requirements.txt
numpy==1.14.2
rx==1.6.1
jsonschema==2.6.0
python-dateutil==2.1
pandas==0.22.0
statsmodels==0.8.0
bokeh==0.12.15
kafka-python==1.4.2
python-ldap==3.0.0
tornado==5.0.2
PyYAML==3.12
oauth2client
google-auth
google-auth-httplib2
google-api-python-client
google-cloud
google-cloud-pubsub
flake8
mock
nose
coverage
cython
findspark
